declutter main and move code to my file
