+++ pft_flutter_hunt/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pft_flutter_hunt/cassidy_m.dart';
+import 'package:pft_flutter_hunt/kris_a.dart';
 
 void main() {
   runApp(const MyApp());
@@ -165,130 +166,15 @@ class _MyHomePageState extends State<MyHomePage> {
                   );
                 },
                 child: const Text('Cassidy M.')),
+              ElevatedButton(
+                onPressed: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(builder: (context) => const KristenPage()),
+                  );
+                },
+                child: const Text('Kristen Averett')),
           ],
-        ),
-      ),
-    );
-  }
-}
-
-class KristenPage extends StatefulWidget {
-  const KristenPage({super.key});
-
-  @override
-  _KristenPageState createState() => _KristenPageState();
-}
-
-class _KristenPageState extends State<KristenPage> {
-  final _formKey = GlobalKey<FormState>();
-  final TextEditingController myController = TextEditingController();
-
-   @override
-  void dispose() {
-    myController.dispose();
-    super.dispose();
-  }
-
-  void _checkInput() {
-    if (myController.text == "2215") {
-      showDialog(
-        context: context,
-        builder: (context) {
-          return const AlertDialog(
-            content: Text("Correct! You've entered 2215."),
-          );
-        },
-      );
-    }
-  }
-
-
-class FirstFloor extends StatelessWidget {
-  const FirstFloor({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('First Floor PFT')),
-      body: Center(
-        child: ElevatedButton(
-          onPressed: () {
-            Navigator.pop(context);
-          },
-          child: const Text('Go Back!'),
-        ),
-      ),
-    );
-  }
-}
-
-class SecondFloor extends StatelessWidget {
-  const SecondFloor({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Second Floor PFT')),
-      body: Center(
-        child: ElevatedButton(
-          onPressed: () {
-            Navigator.pop(context);
-          },
-          child: const Text('Go Back!'),
-        ),
-      ),
-    );
-  }
-}
-
-class ThirdFloor extends StatelessWidget {
-  const ThirdFloor({super.key});
-
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-
-      appBar: AppBar(
-        backgroundColor: const Color.fromARGB(255, 53, 106, 67),
-        title: const Text("Kristen's Page"),
-      ),
-      body: Column(
-        children: <Widget>[
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 16),
-            child: TextFormField(
-              controller: myController,
-              decoration: const InputDecoration(
-                border: UnderlineInputBorder(),
-                labelText: 'Enter the room number!',
-              ),
-            ),
-          ),
-          Expanded(
-            child: Image.asset(
-              'assets/driving_sim.jpeg',
-              fit: BoxFit.cover,
-              width: double.infinity,
-            ),
-          ),
-        ],
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _checkInput, // Call function to check input
-        tooltip: 'Check Number',
-        child: const Icon(Icons.check),
-      ),
-    );
-  }
-}
-      appBar: AppBar(title: const Text('Third Floor PFT')),
-      body: Center(
-        child: ElevatedButton(
-          onPressed: () {
-            Navigator.pop(context);
-          },
-          child: const Text('Go Back!'),
         ),
       ),
     );
